fixed the @media for the frontpage
changed the hight for the picture to be a viewpoint hight so that it stays good compared to the size of the site. removed .frontPage infront of every part of the page, because it was not nessasary. changed @media screen in general.css because it should be (max-width). added a br in the html, because it was the easiest way to devide the text into two different lines.